ダメージ変数の型の不整合を修正

+++ Component/Common/AttackColliderController.h
@@ -28,5 +28,5 @@ public:
 
 private:
     Glib::WeakPtr<Glib::Collider> collider_;
-    float power_{ 0.0f };
+    int power_{ 0 };
 };
